Move JmsInboundGateway to the jms Package
